backend-control-ventas/requirements.txt
alembic==1.15.2
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
cryptography==39.0.1
Flask==3.1.0
flask-cors==5.0.1 
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.40 
Flask-SQLAlchemy==3.1.1 
PyMySQL==1.1.1    
mysql-connector-python==9.2.0 
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
python-dotenv==1.1.0 
typing_extensions==4.13.2
Werkzeug==3.1.3









